cli: Move CLI sources up one level, to src/
