+++ src/type-generator.js
